Major cleanup of code action suppliers
