+++ extension/server/src/providers/linter/index.ts
@@ -6,7 +6,6 @@ import { documents, parser } from '..';
 import { IssueRange, Rules } from '../../../../../language/parserTypes';
 import Linter from '../../../../../language/linter';
 import Cache from '../../../../../language/models/cache';
-import codeActionsProvider from './codeActions';
 import documentFormattingProvider from './documentFormatting';
 
 import * as Project from "../project";
@@ -15,8 +14,11 @@ import { parseMemberUri } from '../../data';
 
 export let jsonCache: { [uri: string]: string } = {};
 
+export function isLinterEnabled() {
+	return true;
+}
+
 export function initialise(connection: _Connection) {
-	connection.onCodeAction(codeActionsProvider);
 	connection.onDocumentFormatting(documentFormattingProvider);
 
 	// This only works for local workspace files
@@ -431,158 +433,4 @@ export function getActions(document: TextDocument, errors: IssueRange[]) {
 	});
 
 	return actions;
-}
-
-export function getSubroutineActions(document: TextDocument, docs: Cache, range: Range): CodeAction|undefined {
-	if (range.start.line === range.end.line) {
-		const currentGlobalSubroutine = docs.subroutines.find(sub => sub.position.range.line === range.start.line && sub.range.start && sub.range.end);
-
-		if (currentGlobalSubroutine) {
-			const subroutineRange = Range.create(
-				Position.create(currentGlobalSubroutine.range.start!, 0),
-				Position.create(currentGlobalSubroutine.range.end!, 1000)
-			);
-
-			const bodyRange = Range.create(
-				Position.create(currentGlobalSubroutine.range.start! + 1, 0),
-				Position.create(currentGlobalSubroutine.range.end! - 1, 0)
-			);
-
-			// First, let's create the extract data
-			const extracted = createExtract(document, bodyRange, docs);
-
-			// Create the new procedure body
-			const newProcedure = [
-				`Dcl-Proc ${currentGlobalSubroutine.name};`,
-				`  Dcl-Pi *N;`,
-					  ...extracted.references.map((ref, i) => `    ${extracted.newParamNames[i]} ${ref.dec.type === `struct` ? `LikeDS` : `Like`}(${ref.dec.name});`),
-				`  End-Pi;`,
-				``,
-				caseInsensitiveReplaceAll(extracted.newBody, `leavesr`, `return`),
-				`End-Proc;`
-			].join(`\n`)
-
-			// Then update the references that invokes this subroutine
-			const referenceUpdates: TextEdit[] = currentGlobalSubroutine.references.map(ref => {
-				const lineNumber = document.positionAt(ref.offset.start).line;
-				// If this reference is outside of the subroutine
-				if (lineNumber < currentGlobalSubroutine.range.start! || lineNumber > currentGlobalSubroutine.range.end!) {
-					return TextEdit.replace(
-						Range.create(
-							// - 5 `EXSR `
-							document.positionAt(ref.offset.start - 5),
-							document.positionAt(ref.offset.end)
-						),
-						currentGlobalSubroutine.name + `(${extracted.references.map(r => r.dec.name).join(`:`)})`
-					);
-				}
-			}).map(x => x) as TextEdit[];
-
-			const refactorAction = CodeAction.create(`Convert to procedure`, CodeActionKind.RefactorExtract);
-			refactorAction.edit = {
-				changes: {
-					[document.uri]: [
-						...referenceUpdates,
-						TextEdit.replace(subroutineRange, newProcedure)
-					]
-				},
-			};
-
-			return refactorAction;
-		}
-	}
-}
-
-export function getExtractProcedureAction(document: TextDocument, docs: Cache, range: Range): CodeAction|undefined {
-	if (range.end.line > range.start.line) {
-			const lastLine = document.offsetAt({line: document.lineCount, character: 0});
-
-			const extracted = createExtract(document, range, docs);
-
-			const newProcedure = [
-				`Dcl-Proc NewProcedure;`,
-				`  Dcl-Pi *N;`,
-					  ...extracted.references.map((ref, i) => `    ${extracted.newParamNames[i]} ${ref.dec.type === `struct` ? `LikeDS` : `Like`}(${ref.dec.name});`),
-				`  End-Pi;`,
-				``,
-				extracted.newBody,
-				`End-Proc;`
-			].join(`\n`)
-
-			const newAction = CodeAction.create(`Extract to new procedure`, CodeActionKind.RefactorExtract);
-
-			// First do the exit
-			newAction.edit = {
-				changes: {
-					[document.uri]: [
-						TextEdit.replace(extracted.range, `NewProcedure(${extracted.references.map(r => r.dec.name).join(`:`)});`),
-						TextEdit.insert(document.positionAt(lastLine), `\n\n`+newProcedure)
-					]
-				},
-			};
-
-			// Then format the document
-			newAction.command = {
-				command: `editor.action.formatDocument`,
-				title: `Format`
-			};
-
-			return newAction;
-	}
-}
-
-function caseInsensitiveReplaceAll(text: string, search: string, replace: string) {
-	return text.replace(new RegExp(search, `gi`), replace);
-}
-	
-
-function createExtract(document: TextDocument, userRange: Range, docs: Cache) {
-	const range = Range.create(userRange.start.line, 0, userRange.end.line, 1000);
-	const references = docs.referencesInRange(document.uri, {start: document.offsetAt(range.start), end: document.offsetAt(range.end)});
-	const validRefs = references.filter(ref => [`struct`, `subitem`, `variable`].includes(ref.dec.type));
-
-	const nameDiffSize = 1; // Always once since we only add 'p' at the start
-	const newParamNames = validRefs.map(ref => `p${ref.dec.name}`);
-	let newBody = document.getText(range);
-
-	const rangeStartOffset = document.offsetAt(range.start);
-
-	// Fix the found offset lengths to be relative to the new procedure
-	for (let i = validRefs.length - 1; i >= 0; i--) {
-		for (let y = validRefs[i].refs.length - 1; y >= 0; y--) {
-			validRefs[i].refs[y] = {
-				start: validRefs[i].refs[y].start - rangeStartOffset,
-				end: validRefs[i].refs[y].end - rangeStartOffset
-			};
-		}
-	}
-
-	// Then let's fix the references to use the new names
-	for (let i = validRefs.length - 1; i >= 0; i--) {
-		for (let y = validRefs[i].refs.length - 1; y >= 0; y--) {
-			const ref = validRefs[i].refs[y];
-
-			newBody = newBody.slice(0, ref.start) + newParamNames[i] + newBody.slice(ref.end);
-			ref.end += nameDiffSize;
-
-			// Then we need to update the offset of the next references
-			for (let z = i - 1; z >= 0; z--) {
-				for (let x = validRefs[z].refs.length - 1; x >= 0; x--) {
-					if (validRefs[z].refs[x].start > ref.end) {
-						validRefs[z].refs[x] = {
-							start: validRefs[z].refs[x].start + nameDiffSize,
-							end: validRefs[z].refs[x].end + nameDiffSize
-						};
-					}
-				}
-			}
-		}
-	}
-
-	return {
-		newBody,
-		newParamNames,
-		references: validRefs,
-		range
-	}
 }
